Return an array of solutions when solving for a given board size
With this change we'll return an array of Solution objects
when solving for a board of size N. A Solution object consists
of a collection of tuples that indicate the positions in which
the queens must be placed so they don't kill each other.

This prepares us to store the solutions in the database, we'll
use an ORM to map the Solution class to a database table.